refactor solve step fns into `solver.rs` file
